+++ src/API.jl
@@ -439,6 +439,8 @@ function test(ctx::Context, pkgs::Vector{PackageSpec};
     return
 end
 
+is_manifest_current(ctx::Context = Context()) = Operations.is_manifest_current(ctx.env)
+
 const UsageDict = Dict{String,DateTime}
 const UsageByDepotDict = Dict{String,UsageDict}
 
@@ -1493,6 +1495,13 @@ function instantiate(ctx::Context; manifest::Union{Bool, Nothing}=nothing,
         pkgerror("expected manifest file at `$(ctx.env.manifest_file)` but it does not exist")
     end
     Types.check_warn_manifest_julia_version_compat(ctx.env.manifest, ctx.env.manifest_file)
+
+    if Operations.is_manifest_current(ctx.env) === false
+        @warn """The project and manifest may be out of sync as either project dependencies have been \
+        added/removed or compat entries have changed since the manifest was last resolved.
+        Try `Pkg.resolve()` or consider `Pkg.update()` if necessary."""
+    end
+
     Operations.prune_manifest(ctx.env)
     for (name, uuid) in ctx.env.project.deps
         get(ctx.env.manifest, uuid, nothing) === nothing || continue