+++ CHANGELOG.md
@@ -10,6 +10,9 @@ Pkg v1.8 Release Notes
   i.e. `pkg> compat Fo<TAB>` autocompletes to `pkg> Foo 0.4,0.5` so that the existing entry can be edited.
 - Pkg now only tries to download packages from the package server in case the
   server tracks a registry that contains the package.
+- `Pkg.instantiate` will now warn when a Project.toml is out of sync with a Manifest.toml. It does this by storing a hash
+  of the project deps and compat entries (other fields are ignored) in the manifest when it is resolved, so that any change
+  to the Project.toml deps or compat entries without a re-resolve can be detected.
 
 Pkg v1.7 Release Notes
 ======================